+++ xllm/core/layers/common/dense_mlp.h
@@ -29,8 +29,8 @@ namespace layer {
 class DenseMLPImpl : public torch::nn::Module {
  public:
   DenseMLPImpl() = default;
-  DenseMLPImpl(int hidden_size,
-               int intermediate_size,
+  DenseMLPImpl(int64_t hidden_size,
+               int64_t intermediate_size,
                bool is_gated,
                bool has_bias,
                const std::string& hidden_act,
