+++ xllm/core/layers/common/layer_utils.cpp
@@ -22,7 +22,7 @@ namespace layer {
 
 bool is_dummy_run(const ModelInputParams& input_params,
                   const ParallelArgs& parallel_args) {
-  int dp_rank = 0;
+  int64_t dp_rank = 0;
   if (parallel_args.dp_size() > 1) {
     dp_rank = parallel_args.dp_local_process_group_->rank();
   }
@@ -32,7 +32,7 @@ bool is_dummy_run(const ModelInputParams& input_params,
   return input_params.dp_global_token_nums[dp_rank] == 0;
 }
 
-void update_dummy_run_input(int dp_rank,
+void update_dummy_run_input(int64_t dp_rank,
                             torch::Tensor& positions,
                             ModelInputParams& input_params) {
   auto& dp_ranks = input_params.dp_global_token_nums;
